fix: correct column descr allocation
There was a bug in the native module where when adding more columns and triggering the more_columndescr_space to allocate more columns, not all the memory from the array and string heaps would get shifted over. This broke column access. Added a test case to hopefully catch this in the future